refactor(rpc): decompose transport and publisher surfaces
Split RPC transport responsibilities into cohesive files, share transaction projection logic, and tighten internal surfaces while preserving wire behavior.

Refs #1483
