+++ internal/rpc/handlers/tx.go
@@ -12,6 +12,7 @@ import (
 
 	binarycodec "github.com/LeJamon/go-xrpl/codec/binarycodec"
 	"github.com/LeJamon/go-xrpl/internal/ledger/service/svcerr"
+	"github.com/LeJamon/go-xrpl/internal/rpc/txprojection"
 	"github.com/LeJamon/go-xrpl/internal/rpc/types"
 	txcore "github.com/LeJamon/go-xrpl/internal/tx"
 	"github.com/LeJamon/go-xrpl/protocol"
@@ -213,7 +214,7 @@ func (m *TxMethod) buildResponseV1(
 			}
 		}
 	} else {
-		maps.Copy(response, projectTransactionJSON(storedTx.TxJSON, "", 1))
+		maps.Copy(response, txprojection.ProjectJSON(storedTx.TxJSON, "", 1))
 		if storedTx.Meta != nil {
 			InjectSyntheticFields(storedTx.TxJSON, storedTx.Meta, SyntheticMetadataContext{
 				LedgerSequence: txInfo.LedgerIndex,
@@ -267,7 +268,7 @@ func (m *TxMethod) buildResponseV2(
 			}
 		}
 	} else {
-		txJSON = projectTransactionJSON(storedTx.TxJSON, "", 2)
+		txJSON = txprojection.ProjectJSON(storedTx.TxJSON, "", 2)
 		// date and ledger_index go inside tx_json for v2
 		if txInfo.LedgerIndex > 0 {
 			txJSON["ledger_index"] = txInfo.LedgerIndex
