+++ internal/rpc/publisher.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"strconv"
 
-	"github.com/LeJamon/go-xrpl/internal/rpc/handlers"
 	"github.com/LeJamon/go-xrpl/internal/rpc/subscription"
 	"github.com/LeJamon/go-xrpl/internal/rpc/types"
 	xrpllog "github.com/LeJamon/go-xrpl/log"
@@ -72,49 +71,6 @@ func (p *Publisher) PublishLedgerClosed(event *LedgerCloseEvent) {
 	}
 
 	p.manager.BroadcastToStream(types.SubLedger, data, nil)
-}
-
-// PublishTransaction broadcasts a transaction event to subscribers
-func (p *Publisher) PublishTransaction(event *TransactionEvent, affectedAccounts []string) {
-	if event == nil || p.manager == nil {
-		return
-	}
-
-	v1, err := marshalTransactionEvent(event, types.ApiVersion1)
-	if err != nil {
-		xrpllog.Named(xrpllog.PartitionRPC).Error("Failed to marshal TransactionEvent", "err", err)
-		return
-	}
-	v2, err := marshalTransactionEvent(event, types.ApiVersion2)
-	if err != nil {
-		xrpllog.Named(xrpllog.PartitionRPC).Error("Failed to marshal TransactionEvent", "err", err)
-		return
-	}
-
-	p.manager.BroadcastToStreamVersioned(types.SubTransactions, v1, v2)
-	p.manager.BroadcastToStreamVersioned(types.SubTransactionsProposed, v1, v2)
-
-	if len(affectedAccounts) > 0 {
-		p.manager.BroadcastToAcceptedAccountsVersioned(v1, v2, affectedAccounts)
-	}
-}
-
-func marshalTransactionEvent(event *TransactionEvent, apiVersion int) ([]byte, error) {
-	txJSON, err := handlers.ProjectTransactionRaw(event.Transaction, event.Hash, apiVersion)
-	if err != nil {
-		return nil, err
-	}
-
-	projected := *event
-	if apiVersion > 1 {
-		projected.Transaction = nil
-		projected.TxJson = txJSON
-	} else {
-		projected.Transaction = txJSON
-		projected.TxJson = nil
-		projected.Hash = ""
-	}
-	return json.Marshal(&projected)
 }
 
 // PublishValidation broadcasts a validation event to validation stream subscribers
@@ -212,67 +168,6 @@ func (p *Publisher) PublishPeerStatus(event *PeerStatusEvent) {
 	p.manager.BroadcastToStream(types.SubPeerStatus, data, nil)
 }
 
-// PublishProposedTransaction broadcasts a proposed transaction to accounts_proposed subscribers
-func (p *Publisher) PublishProposedTransaction(event *ProposedTransactionEvent, accounts []string) {
-	if event == nil || p.manager == nil {
-		return
-	}
-
-	v1, err := marshalProposedTransactionEvent(event, types.ApiVersion1)
-	if err != nil {
-		xrpllog.Named(xrpllog.PartitionRPC).Error("Failed to marshal ProposedTransactionEvent", "err", err)
-		return
-	}
-	v2, err := marshalProposedTransactionEvent(event, types.ApiVersion2)
-	if err != nil {
-		xrpllog.Named(xrpllog.PartitionRPC).Error("Failed to marshal ProposedTransactionEvent", "err", err)
-		return
-	}
-
-	p.manager.BroadcastToStreamVersioned(types.SubTransactionsProposed, v1, v2)
-	if len(accounts) > 0 {
-		p.manager.BroadcastToAccountsProposedVersioned(v1, v2, accounts)
-	}
-}
-
-func marshalProposedTransactionEvent(event *ProposedTransactionEvent, apiVersion int) ([]byte, error) {
-	txJSON, err := handlers.ProjectTransactionRaw(event.Transaction, event.Hash, apiVersion)
-	if err != nil {
-		return nil, err
-	}
-
-	projected := *event
-	if apiVersion > 1 {
-		projected.Transaction = nil
-		projected.TxJson = txJSON
-	} else {
-		projected.Transaction = txJSON
-		projected.TxJson = nil
-		projected.Hash = ""
-	}
-	return json.Marshal(&projected)
-}
-
-// PublishOrderBookChange broadcasts an order book change to book subscribers
-func (p *Publisher) PublishOrderBookChange(event *TransactionEvent, books []types.OrderBookSpec) {
-	if event == nil || p.manager == nil || len(books) == 0 {
-		return
-	}
-
-	v1, err := marshalTransactionEvent(event, types.ApiVersion1)
-	if err != nil {
-		xrpllog.Named(xrpllog.PartitionRPC).Error("Failed to marshal order-book TransactionEvent", "err", err)
-		return
-	}
-	v2, err := marshalTransactionEvent(event, types.ApiVersion2)
-	if err != nil {
-		xrpllog.Named(xrpllog.PartitionRPC).Error("Failed to marshal order-book TransactionEvent", "err", err)
-		return
-	}
-
-	p.manager.BroadcastToOrderBooksVersioned(v1, v2, books)
-}
-
 // GetSubscriberCount returns the number of active subscribers for a stream type
 func (p *Publisher) GetSubscriberCount(streamType types.SubscriptionType) int {
 	if p.manager == nil {
@@ -281,26 +176,5 @@ func (p *Publisher) GetSubscriberCount(streamType types.SubscriptionType) int {
 	return p.manager.GetSubscriberCount(streamType)
 }
 
-// NoOpPublisher is a publisher that does nothing (for testing or when subscriptions are disabled)
-type NoOpPublisher struct{}
-
-func NewNoOpPublisher() *NoOpPublisher {
-	return &NoOpPublisher{}
-}
-
-func (p *NoOpPublisher) PublishLedgerClosed(event *LedgerCloseEvent)                   {}
-func (p *NoOpPublisher) PublishTransaction(event *TransactionEvent, accounts []string) {}
-func (p *NoOpPublisher) PublishValidation(event *ValidationEvent)                      {}
-func (p *NoOpPublisher) PublishServerStatus(event *ServerStatusEvent) bool             { return false }
-func (p *NoOpPublisher) PublishConsensusPhase(phase string)                            {}
-func (p *NoOpPublisher) PublishManifest(event *ManifestEvent)                          {}
-func (p *NoOpPublisher) PublishPeerStatus(event *PeerStatusEvent)                      {}
-func (p *NoOpPublisher) PublishProposedTransaction(event *ProposedTransactionEvent, accounts []string) {
-}
-func (p *NoOpPublisher) PublishOrderBookChange(event *TransactionEvent, books []types.OrderBookSpec) {
-}
-func (p *NoOpPublisher) GetSubscriberCount(streamType types.SubscriptionType) int { return 0 }
-
 // Ensure implementations satisfy the interface
 var _ EventPublisher = (*Publisher)(nil)
-var _ EventPublisher = (*NoOpPublisher)(nil)
